requirements.txt
discord
dotenv
request